requirements.txt
Adafruit-Blinka==8.28.1
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-framebuf==1.6.5
adafruit-circuitpython-requests==2.0.3
adafruit-circuitpython-ssd1306==2.12.16
adafruit-circuitpython-typing==1.9.6
Adafruit-PlatformDetect==3.57.0
Adafruit-PureIO==1.1.11
adafruit-python-shell==1.8.0
annotated-types==0.6.0
anyio==4.2.0
args==0.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
clint==0.5.1
distro==1.9.0
exceptiongroup==1.2.0
Flask==3.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
openai==1.6.1
Pillow==10.1.0
pydantic==2.5.3
pydantic_core==2.14.6
pyftdi==0.55.0
pyserial==3.5
pyusb==1.2.1
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
